Use process_vm_readv/writev instead of proc
Changing the to the target mount namespace, then the pid of the target
process doesn't match anymore. E.g. if the target process is the
container entrypoint, then its pid will be 1 and we should access its
memory through /proc/1/mem and not from /proc/<pid form request>/mem.

In order to avoid this mapping, we use instead process_vm_readv/writev
to read and write to the target process memory. In this way, we avoid
the issue of the changing mount namespace, plus we use a single syscall
instead of opening the file and the pread64/pwrite syscalls.

Signed-off-by: Alice Frosi <afrosi@redhat.com>